# packages/docs/src/openapi/eliza-v1.yaml
openapi: 3.0.0
info:
  title: alan OS API
  description: |-
    API documentation for alan OS v1.0.10 - A flexible and scalable AI agent framework.

    This API is designed to be used with a locally running alan instance. Endpoints allow for creating,
    managing, and interacting with AI agents through a REST interface.

    The API is organized into the following domains:
    - **System**: System-wide operations and environment management
    - **Agents**: Agent lifecycle and management operations
    - **Memory**: Agent memory and room management
    - **Messaging**: Message handling, channels, and servers
    - **Audio**: Audio processing and speech synthesis
    - **Media**: File upload and media management
    - **TEE**: Trusted Execution Environment operations
    - **WebSocket**: Real-time communication via Socket.IO
  version: 1.0.10
  contact:
    name: alan OS Community
    url: https://github.com/alanos/alan
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/agents/worlds:
    get:
      tags:
        - agents
      summary: Get all worlds
      description: Get all worlds across all agents
      operationId: getAllWorlds
      responses:
        '200':
          description: Worlds retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      worlds:
                        type: array
                        items:
                          $ref: '#/components/schemas/World'
        '500':
          description: Error retrieving worlds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/agents/{agentId}/rooms/{roomId}:
    get:
      tags:
        - rooms
        - agents
      summary: Get room details
      description: Retrieves details about a specific room
      operationId: getRoom
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the room
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Room'
        '400':
          description: Invalid agent ID or room ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Agent or room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error retrieving room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - rooms
        - agents
      summary: Update a room
      description: Updates a specific room
      operationId: updateRoom
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the room to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: New name for the room
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Room'
        '400':
          description: Invalid agent ID or room ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Agent or room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error updating room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - rooms
        - agents
      summary: Delete a room
      description: Deletes a specific room
      operationId: deleteRoom
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the room to delete
      responses:
        '204':
          description: Room deleted successfully
        '400':
          description: Invalid agent ID or room ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Agent or room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error deleting room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/memory/{agentId}/rooms:
    post:
      tags:
        - memory
      summary: Create a room
      description: Create a new room for an agent
      operationId: createMemoryRoom
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Name of the room
                type:
                  type: string
                  enum: [DM, GROUP, CHANNEL]
                  default: DM
                  description: Type of room
                source:
                  type: string
                  default: client
                  description: Source of the room
                worldId:
                  type: string
                  format: uuid
                  description: ID of the world
                serverId:
                  type: string
                  description: Server ID
                metadata:
                  type: object
                  description: Additional room metadata
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      agentId:
                        type: string
                        format: uuid
                      createdAt:
                        type: integer
                      source:
                        type: string
                      type:
                        type: string
                      worldId:
                        type: string
                        format: uuid
                      serverId:
                        type: string
                      metadata:
                        type: object
        '400':
          description: Invalid agent ID or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error creating room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Messaging endpoints
...
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
            details:
              type: string
              description: Detailed error information
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the room
        name:
          type: string
          description: Name of the room
        source:
          type: string
          description: Source of the room
        worldId:
          type: string
          format: uuid
          description: ID of the world this room belongs to
        entities:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
          description: Entities in this room
...
    World:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the world
        name:
          type: string
          description: Name of the world
        agentId:
          type: string
          format: uuid
          description: ID of the agent that owns this world
        sourceType:
          type: string
          description: Type of source (discord, telegram, etc)
        sourceId:
          type: string
          description: Platform-specific identifier
        metadata:
          type: object
          description: Additional world metadata
...